+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/alexdlaird/pyngrok/compare/5.0.4...HEAD)
+### Fixed
+- Build improvements.
 
 ## [5.0.4](https://github.com/alexdlaird/pyngrok/compare/5.0.3...5.0.4) - 2021-03-08
 ### Fixed